change: data classをcoreモジュール内に移動
