+++ .idea/deploymentTargetSelector.xml
@@ -8,6 +8,9 @@
       <SelectionState runConfigName="model">
         <option name="selectionMode" value="DROPDOWN" />
       </SelectionState>
+      <SelectionState runConfigName="core">
+        <option name="selectionMode" value="DROPDOWN" />
+      </SelectionState>
     </selectionStates>
   </component>
 </project>
