+++ app/src/main/java/com/example/taskmate/ui/setting/componets/SettingCard.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.taskmate.ui.setting.SettingItem
+import com.example.core.model.SettingItem
 
 @Composable
 fun SettingCard(item: SettingItem) {
