+++ app/src/main/java/com/example/taskmate/ui/task/TaskCard.kt
@@ -15,6 +15,7 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.example.core.model.Task
 import com.example.taskmate.ui.task.components.DisplayDate
 
 @Composable
