+++ core/src/main/java/com/example/core/model/Class.kt
@@ -1,0 +1,6 @@
+package com.example.core.model
+
+data class Class(
+    val day: String,
+    val classList: List<String>,
+)
