+++ core/src/main/java/com/example/core/model/SettingItem.kt
@@ -1,0 +1,8 @@
+package com.example.core.model
+
+data class SettingItem(
+    val title: String,
+    val subTitle: String = "",
+    val icon: Int,
+    val onClick: () -> Unit = {},
+)
