+++ core/src/main/java/com/example/core/model/Tag.kt
@@ -2,7 +2,7 @@ package com.example.core.model
 
 import androidx.compose.ui.graphics.Color
 
-data class TaskmateSubject(
+data class Tag(
     val name: String,
     val color: Color,
 )
