+++ core/src/main/java/com/example/core/model/Task.kt
@@ -1,0 +1,9 @@
+package com.example.core.model
+
+data class Task(
+    val title: String,
+//    val destination: String,
+//    val deadlineDate: Date,
+//    val deadlineTime: Time,
+//    val remind: Int,
+)
