+++ core/src/test/java/com/example/core/ExampleUnitTest.kt
@@ -1,8 +1,8 @@
 package com.example.core
 
+import junit.framework.TestCase.assertEquals
 import org.junit.Test
 
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
